dcons

+++ src/header/ClauseBase.hpp
@@ -159,11 +159,15 @@ public:
 
   void unit_propagate(const Literal<T> l);
 
+  void clearTriggers();
+
   void assign(const Literal<T> l, const NewExplanation<T> &e);
 
   std::ostream &display(std::ostream &os) const;
+  std::ostream &displayWatchStruct(std::ostream &os) const;
 
-//  std::ostream &displayClause(std::ostream &os, const NewClause<T> *c) const;
+  //  std::ostream &displayClause(std::ostream &os, const NewClause<T> *c)
+  //  const;
 
   bool satisfied(const Literal<T>) const;
   bool falsified(const Literal<T>) const;
@@ -189,6 +193,15 @@ public:
   double inverseActivity(const Literal<T> l);
   double loosenessOverActivity(const Literal<T> l);
 
+  NewClause<T> *operator[](const index_t i) {
+
+    //        std::cout << i << std::endl << free_cl_indices << std::endl;
+
+    if (i >= free_cl_indices.capacity() or free_cl_indices.has(i))
+      return NULL;
+    return base[i];
+  }
+
   NewClause<T> *back() {
     //
     //    std::cout << base.size() << " / "
@@ -207,35 +220,36 @@ public:
     virtual void propagate();
     // notify a change (with the literal and it's variable rank in the scope)
     virtual bool notify(const Literal<T>, const int);
-    
-    
-private:
-  Solver<T> &caller;
 
-  SubscriberHandle handlerToken;
+    NewClause<T> *consistent();
 
-  /// Clause memory store ///
-  /// The set of clauses
-  std::vector<NewClause<T> *> base;
-  std::vector<double> score;
-  /// free indices in the vector base (so that we can remove clauses)
-  SparseSet<int> free_cl_indices;
-  // clauses, watch struct watch[BOUND_LIT] for bound literals,
-  // watch[EDGE_LIT] for edges
-  std::vector<std::vector<NewClause<T> *>> watch[2];
-  ////////////////////////////////////////
+  private:
+    Solver<T> &solver;
 
-  //    /// Literal memory store ///
-  //    // store the set of bound constraints appearing in a literal
-  //    std::vector<BoundConstraint<T>> constraint;
-  //    // the number of clauses in which this bound-constraint appears
-  //    std::vector<int> cardinality;
-  //    // free indices in the vector constraints (so that we can remove
-  //    constraints when they are not used) SparseSet<int> free_lit_indices;
-  //    // for every event-lit, the set of (pointers to) constraints ordered by
-  //    bound std::vector<std::vector<int>> cons_list;
-  //    ////////////////////////////////////////
-    
+    SubscriberHandle handlerToken;
+
+    /// Clause memory store ///
+    /// The set of clauses
+    std::vector<NewClause<T> *> base;
+    std::vector<double> score;
+    /// free indices in the vector base (so that we can remove clauses)
+    SparseSet<int> free_cl_indices;
+    // clauses, watch struct watch[BOUND_LIT] for bound literals,
+    // watch[EDGE_LIT] for edges
+    std::vector<std::vector<NewClause<T> *>> watch[2];
+    ////////////////////////////////////////
+
+    //    /// Literal memory store ///
+    //    // store the set of bound constraints appearing in a literal
+    //    std::vector<BoundConstraint<T>> constraint;
+    //    // the number of clauses in which this bound-constraint appears
+    //    std::vector<int> cardinality;
+    //    // free indices in the vector constraints (so that we can remove
+    //    constraints when they are not used) SparseSet<int> free_lit_indices;
+    //    // for every event-lit, the set of (pointers to) constraints ordered
+    //    by bound std::vector<std::vector<int>> cons_list;
+    //    ////////////////////////////////////////
+
     SparseSet<var_t> triggered_bounds;
 
   // statistics
@@ -1342,7 +1356,7 @@ std::ostream &operator<<(std::ostream &os, const tempo::ClauseBase<T> &x) {
 
 template <typename T>
 NewClauseBase<T>::NewClauseBase(Solver<T> &c)
-    : caller(c), handlerToken(caller.SearchRestarted.subscribe_handled(
+    : solver(c), handlerToken(solver.SearchRestarted.subscribe_handled(
                      [this]() { this->forget(); })) {
         
         NewConstraint<T>::priority = Priority::Low;
@@ -1353,26 +1367,33 @@ NewClauseBase<T>::NewClauseBase(Solver<T> &c)
 template <typename T>
 void NewClauseBase<T>::post(const int idx) {
     NewConstraint<T>::cons_id = idx;
-//    for(var_t x{0}; x<caller.boolean.size(); ++x) {
-//        caller.wake_me_on(caller.boolean.getLiteral(true, x), NewConstraint<T>::cons_id);
-//        caller.wake_me_on(caller.boolean.getLiteral(false, x), NewConstraint<T>::cons_id);
+//    for(var_t x{0}; x<solver.boolean.size(); ++x) {
+//        solver.wake_me_on(solver.boolean.getLiteral(true, x), NewConstraint<T>::cons_id);
+//        solver.wake_me_on(solver.boolean.getLiteral(false, x), NewConstraint<T>::cons_id);
 //    }
     
-//        for(var_t x{0}; x<caller.numeric.size(); ++x) {
-//            caller.wake_me_on(lb<T>(x), NewConstraint<T>::cons_id);
-//            caller.wake_me_on(ub<T>(x), NewConstraint<T>::cons_id);
+//        for(var_t x{0}; x<solver.numeric.size(); ++x) {
+//            solver.wake_me_on(lb<T>(x), NewConstraint<T>::cons_id);
+//            solver.wake_me_on(ub<T>(x), NewConstraint<T>::cons_id);
 //        }
-    triggered_bounds.reserve(2 * caller.numeric.size());
+    triggered_bounds.reserve(2 * solver.numeric.size());
 }
 
 // propagate the constraint
 template <typename T>
 void NewClauseBase<T>::propagate() {
     for(auto b : triggered_bounds) {
-        unit_propagate(caller.numeric.strongestLiteral(Literal<T>::sign(b), Literal<T>::var(b)));
+      auto p{solver.numeric.strongestLiteral(Literal<T>::sign(b),
+                                             Literal<T>::var(b))};
+//      std::cout << " --> unitprop " << solver.pretty(p) << std::endl;
+      unit_propagate(p);
     }
-    triggered_bounds.clear();
-//    std::cout << triggered_bounds << std::endl;
+    clearTriggers();
+    //    std::cout << triggered_bounds << std::endl;
+}
+
+template <typename T> void NewClauseBase<T>::clearTriggers() {
+  triggered_bounds.clear();
 }
 
 template <typename T>
@@ -1425,24 +1446,89 @@ template <typename T> void NewClauseBase<T>::newNumericVar(const var_t x) {
 template <typename T>
 bool NewClauseBase<T>::satisfied(const Literal<T> l) const {
   if (l.isNumeric())
-    return caller.numeric.satisfied(l);
+    return solver.numeric.satisfied(l);
   else
-    return caller.boolean.satisfied(l);
+    return solver.boolean.satisfied(l);
 }
 
 template <typename T>
 bool NewClauseBase<T>::falsified(const Literal<T> l) const {
   if (l.isNumeric())
-    return caller.numeric.falsified(l);
+    return solver.numeric.falsified(l);
   else
-    return caller.boolean.falsified(l);
+    return solver.boolean.falsified(l);
 }
 
 // template <typename T>
 // void NewClauseBase<T>::assign(const Literal<T> l, const NewExplanation<T> &e)
 // {
-//   caller.set(l, e);
+//   solver.set(l, e);
 // }
+
+template <typename T> NewClause<T> *NewClauseBase<T>::consistent() {
+  //    for(var_t x{0}; x<solver.boolean.size(); ++x) {
+  //        for(auto cl : watch[BOOLEAN][Literal<T>(true,x)]) {
+  //            assert
+  //        }
+  //    }
+
+  for (auto i{free_cl_indices.fbegin()}; i != free_cl_indices.fend(); ++i) {
+    auto cl{base[*i]};
+    auto w0{cl->watched(0)};
+    auto w1{cl->watched(1)};
+    if (satisfied(w0))
+      continue;
+    if (satisfied(w1))
+      continue;
+    if (falsified(w0)) {
+      if (not satisfied(w1))
+        return cl;
+      for (auto p : *cl) {
+        if (p != w0 and p != w1 and not falsified(p)) {
+          return cl;
+        }
+      }
+    } else if (falsified(w1)) {
+      if (not satisfied(w0))
+        return cl;
+      for (auto p : *cl) {
+        if (p != w0 and p != w1 and not falsified(p)) {
+          return cl;
+        }
+      }
+    }
+  }
+
+  for (auto i{free_cl_indices.bbegin()}; i != free_cl_indices.bend(); ++i) {
+    auto cl{base[*i]};
+    auto w0{cl->watched(0)};
+    auto w1{cl->watched(1)};
+    if (satisfied(w0))
+      continue;
+    if (satisfied(w1))
+      continue;
+    if (falsified(w0)) {
+      if (not satisfied(w1))
+        return cl;
+      for (auto p : *cl) {
+        if (p != w0 and p != w1 and not falsified(p)) {
+          return cl;
+        }
+      }
+    } else if (falsified(w1)) {
+      if (not satisfied(w0))
+        return cl;
+      for (auto p : *cl) {
+        if (p != w0 and p != w1 and not falsified(p)) {
+          return cl;
+        }
+      }
+    }
+  }
+
+  //    std::cout << "ok\n";
+  return NULL;
+}
 
 template <typename T>
 void NewClauseBase<T>::unit_propagate(const Literal<T> l) {
@@ -1452,7 +1538,7 @@ void NewClauseBase<T>::unit_propagate(const Literal<T> l) {
 #endif
 
 #ifdef DBG_TRACE
-  if (DBG_TRACE & UNITPROPAGATION) {
+  if (DBG_CBOUND and (DBG_TRACE & UNITPROPAGATION)) {
     std::cout << "unit propagate true lit " << l << "\n";
   }
 #endif
@@ -1468,7 +1554,7 @@ void NewClauseBase<T>::unit_propagate(const Literal<T> l) {
     auto cl{*c};
 
 #ifdef DBG_TRACE
-    if (DBG_TRACE & UNITPROPAGATION) {
+    if (DBG_CBOUND and (DBG_TRACE & UNITPROPAGATION)) {
       std::cout << " watched by " << *cl << std::endl;
     }
 #endif
@@ -1495,7 +1581,7 @@ void NewClauseBase<T>::unit_propagate(const Literal<T> l) {
           if(c_lit.value() + l.value() >= 0) {
               
 #ifdef DBG_TRACE
-              if (DBG_TRACE & UNITPROPAGATION) {
+              if (DBG_CBOUND and (DBG_TRACE & UNITPROPAGATION)) {
                   std::cout << " false trigger (" << c_lit << " is not falsified)\n";
               }
 #endif
@@ -1504,7 +1590,7 @@ void NewClauseBase<T>::unit_propagate(const Literal<T> l) {
           } else {
               
 #ifdef DBG_TRACE
-              if (DBG_TRACE & UNITPROPAGATION) {
+              if (DBG_CBOUND and (DBG_TRACE & UNITPROPAGATION)) {
                   std::cout << " true trigger (" << c_lit << " and " << l << " are contradictory)\n";
               }
 #endif
@@ -1515,7 +1601,7 @@ void NewClauseBase<T>::unit_propagate(const Literal<T> l) {
     if (satisfied(other)) {
 
 #ifdef DBG_TRACE
-      if (DBG_TRACE & UNITPROPAGATION) {
+      if (DBG_CBOUND and (DBG_TRACE & UNITPROPAGATION)) {
         std::cout << ": satisfied by " << other << std::endl;
       }
 #endif
@@ -1526,7 +1612,7 @@ void NewClauseBase<T>::unit_propagate(const Literal<T> l) {
     index_t i{idx};
 
 #ifdef DBG_TRACE
-      if (DBG_TRACE & UNITPROPAGATION) {
+      if (DBG_CBOUND and (DBG_TRACE & UNITPROPAGATION)) {
         std::cout << "search another literal to watch" ;
       }
 #endif
@@ -1541,7 +1627,7 @@ void NewClauseBase<T>::unit_propagate(const Literal<T> l) {
       auto p = (*cl)[i];
 
 #ifdef DBG_TRACE
-      if (DBG_TRACE & UNITPROPAGATION) {
+      if (DBG_CBOUND and (DBG_TRACE & UNITPROPAGATION)) {
         std::cout << "  " << p;
       }
 #endif
@@ -1550,7 +1636,7 @@ void NewClauseBase<T>::unit_propagate(const Literal<T> l) {
         if (not falsified(p)) {
 
 #ifdef DBG_TRACE
-          if (DBG_TRACE & UNITPROPAGATION) {
+          if (DBG_CBOUND and (DBG_TRACE & UNITPROPAGATION)) {
             std::cout << ": replace by " << p << " (" << i << ") as "
                       << watch_rank << "-th watcher ";
           }
@@ -1559,7 +1645,7 @@ void NewClauseBase<T>::unit_propagate(const Literal<T> l) {
           set_watcher(watch_rank, i, cl);
 
 #ifdef DBG_TRACE
-          if (DBG_TRACE & UNITPROPAGATION) {
+          if (DBG_CBOUND and (DBG_TRACE & UNITPROPAGATION)) {
             std::cout << " and rm from " << l << "'s watches\n";
           }
 #endif
@@ -1572,7 +1658,7 @@ void NewClauseBase<T>::unit_propagate(const Literal<T> l) {
         }
       }
 #ifdef DBG_TRACE
-      else if (DBG_TRACE & UNITPROPAGATION) {
+      else if (DBG_CBOUND and (DBG_TRACE & UNITPROPAGATION)) {
         std::cout << "*";
       }
 #endif
@@ -1581,7 +1667,7 @@ void NewClauseBase<T>::unit_propagate(const Literal<T> l) {
     if (i == idx) {
 
 #ifdef DBG_TRACE
-      if (DBG_TRACE & UNITPROPAGATION) {
+      if (DBG_CBOUND and (DBG_TRACE & UNITPROPAGATION)) {
         std::cout << ": new unit " << other << std::endl;
       }
 #endif
@@ -1590,7 +1676,7 @@ void NewClauseBase<T>::unit_propagate(const Literal<T> l) {
       verifyWatchers("at assign");
 #endif
 
-      caller.set(other, {this, cl->id});
+      solver.set(other, {this, cl->id});
 
       assert(cl == base[cl->id]);
     }
@@ -1659,16 +1745,16 @@ template <typename T> void NewClauseBase<T>::forget_worst() {
 
 template <typename T>
 double NewClauseBase<T>::loosenessOverActivity(const Literal<T> l) {
-  return caller.looseness(l) / caller.getActivityMap()->get(l, caller);
+  return solver.looseness(l) / solver.getActivityMap()->get(l, solver);
 }
 
 template <typename T>
 double NewClauseBase<T>::inverseActivity(const Literal<T> l) {
-  return 1.0 / caller.getActivityMap()->get(l, caller);
+  return 1.0 / solver.getActivityMap()->get(l, solver);
 }
 
 template <typename T> double NewClauseBase<T>::looseness(const Literal<T> l) {
-  return caller.looseness(l);
+  return solver.looseness(l);
 }
 
 template <typename T> void NewClauseBase<T>::forgetAll() {
@@ -1686,9 +1772,9 @@ template <typename T> void NewClauseBase<T>::forget() {
   verifyWatchers("before forget");
 #endif
 
-  if (caller.getOptions().forget_strategy == Options::LiteralScore::Looseness or
-      ( // caller.getActivityMap() == NULL and
-          caller.getOptions().forget_strategy != Options::LiteralScore::Size)) {
+  if (solver.getOptions().forget_strategy == Options::LiteralScore::Looseness or
+      ( // solver.getActivityMap() == NULL and
+          solver.getOptions().forget_strategy != Options::LiteralScore::Size)) {
     for (auto idx{free_cl_indices.bbegin()}; idx != free_cl_indices.bend();
          ++idx) {
       score[*idx] = 0;
@@ -1699,7 +1785,7 @@ template <typename T> void NewClauseBase<T>::forget() {
         score[*idx] += looseness(l);
       }
     }
-  } else if (caller.getOptions().forget_strategy ==
+  } else if (solver.getOptions().forget_strategy ==
              Options::LiteralScore::LoosenessOverActivity) {
     for (auto idx{free_cl_indices.bbegin()}; idx != free_cl_indices.bend();
          ++idx) {
@@ -1711,7 +1797,7 @@ template <typename T> void NewClauseBase<T>::forget() {
         score[*idx] += loosenessOverActivity(l);
       }
     }
-  } else if (caller.getOptions().forget_strategy ==
+  } else if (solver.getOptions().forget_strategy ==
              Options::LiteralScore::Activity) {
     for (auto idx{free_cl_indices.bbegin()}; idx != free_cl_indices.bend();
          ++idx) {
@@ -1725,7 +1811,7 @@ template <typename T> void NewClauseBase<T>::forget() {
     }
   }
 
-  if (caller.getOptions().forget_strategy == Options::LiteralScore::Size) {
+  if (solver.getOptions().forget_strategy == Options::LiteralScore::Size) {
     std::sort(free_cl_indices.bbegin(), free_cl_indices.bend(),
               [&](const int i, const int j) {
                 return (base[i]->size() > base[j]->size());
@@ -1740,7 +1826,7 @@ template <typename T> void NewClauseBase<T>::forget() {
   free_cl_indices.re_index(free_cl_indices.bbegin(), free_cl_indices.bend());
 
   auto target_size = static_cast<size_t>(
-      static_cast<double>(numLearnt()) * (1.0 - caller.getOptions().forgetfulness));
+      static_cast<double>(numLearnt()) * (1.0 - solver.getOptions().forgetfulness));
 
   while (numLearnt() > target_size) {
     forget_worst();
@@ -1772,7 +1858,7 @@ NewClause<T> *NewClauseBase<T>::add(const iter first, const iter last,
 
   if (first + 1 == last) {
     //    assign(*first, Constant::NewNoReason<T>);
-    caller.set(*first, Constant::GroundFact<T>);
+    solver.set(*first, Constant::GroundFact<T>);
   } else {
     if (not free_cl_indices.empty()) {
       int id{free_cl_indices.back()};
@@ -1816,7 +1902,7 @@ NewClause<T> *NewClauseBase<T>::add(const iter first, const iter last,
       if(learnt) {
           Literal<T> l{(*c)[0]};
           //          assign(l, {this, c->id});
-          caller.set(l, {this, c->id});
+          solver.set(l, {this, c->id});
       }
   }
 
@@ -1924,7 +2010,13 @@ template <typename T> int NewClauseBase<T>::getType() const {
 }
 
 template <typename T>
-std::ostream &NewClauseBase<T>::display(
+std::ostream &NewClauseBase<T>::display(std::ostream &os) const {
+  os << "clause base";
+  return os;
+}
+
+template <typename T>
+std::ostream &NewClauseBase<T>::displayWatchStruct(
     std::ostream &os
     //                                            , const std::function<int>&
     //                                            f=TODIMACS
@@ -1936,8 +2028,8 @@ std::ostream &NewClauseBase<T>::display(
   //        else
   //            os << "(deleted clause " << cl->id << ")\n";
 
-  for (size_t x{0}; x < caller.boolean.size(); ++x) {
-    auto l{caller.boolean.getLiteral(true, x)};
+  for (size_t x{0}; x < solver.boolean.size(); ++x) {
+    auto l{solver.boolean.getLiteral(true, x)};
     if (not watch[BOOLEAN][l].empty()) {
       os << l << " is watched in";
       for (auto cl : watch[BOOLEAN][l]) {
@@ -1953,7 +2045,7 @@ std::ostream &NewClauseBase<T>::display(
       os << std::endl;
     }
   }
-  for (size_t x{0}; x < caller.numeric.size(); ++x) {
+  for (size_t x{0}; x < solver.numeric.size(); ++x) {
     auto l = lb<T>(x);
     if (not watch[NUMERIC][l].empty()) {
       os << l << " is watched in";
@@ -1990,8 +2082,8 @@ void NewClauseBase<T>::verifyWatchers(const char *msg) const {
     }
     
     size_t num_watchers{0};
-    for (size_t x{0}; x < caller.boolean.size(); ++x) {
-        Literal<T> l{caller.boolean.getLiteral(true,x)};
+    for (size_t x{0}; x < solver.boolean.size(); ++x) {
+        Literal<T> l{solver.boolean.getLiteral(true,x)};
         num_watchers += watch[BOOLEAN][l].size();
         if (not watch[BOOLEAN][l].empty()) {
             for (auto cl : watch[BOOLEAN][l]) {
@@ -2036,7 +2128,7 @@ void NewClauseBase<T>::verifyWatchers(const char *msg) const {
             }
         }
     }
-    for (size_t x{0}; x < caller.numeric.size(); ++x) {
+    for (size_t x{0}; x < solver.numeric.size(); ++x) {
       Literal<T> l{lb<T>(x)};
       num_watchers += watch[NUMERIC][l].size();
       if (not watch[NUMERIC][l].empty()) {
